feat: Moved and merged providers
