+++ src/app/(routes)/layout.tsx
@@ -1,7 +1,7 @@
 import '@/app/globals.css'
 import { SidebarProvider } from '@/components/sidebar/primitive'
-import { OrganizationProvider } from '@/context/organization-provider'
-import { PermissionProvider } from '@/context/permission-provider'
+import { OrganizationProvider } from '@/providers/organization-provider'
+import { PermissionProvider } from '@/providers/permission-provider'
 import { nextAuthOptions } from '@/core/infrastructure/next-auth/next-auth-provider'
 import { getServerSession } from 'next-auth'
 import { redirect, RedirectType } from 'next/navigation'
