+++ src/app/(routes)/transactions/create/basic-information-paper.tsx
@@ -3,7 +3,7 @@ import { InputField, SelectField } from '@/components/form'
 import { Paper } from '@/components/ui/paper'
 import { SelectItem } from '@/components/ui/select'
 import { Separator } from '@/components/ui/separator'
-import { useOrganization } from '@/context/organization-provider/organization-provider-client'
+import { useOrganization } from '@/providers/organization-provider/organization-provider-client'
 import { Control } from 'react-hook-form'
 import { useIntl } from 'react-intl'
 import DolarSign from '/public/svg/dolar-sign.svg'
