+++ src/components/organization-switcher/index.tsx
@@ -8,6 +8,7 @@ import React, { useEffect } from 'react'
 import { useIntl } from 'react-intl'
 import { useSidebar } from '../sidebar/primitive'
 import { Skeleton } from '../ui/skeleton'
+import { useOrganization } from '@/providers/organization-provider/organization-provider-client'
 import { OrganizationSwitcherContent } from './organization-switcher-content'
 import { SwitcherTrigger } from './organization-switcher-trigger'
 import MidazLogo from '/public/svg/brand-midaz.svg'
