+++ src/providers/permission-provider/validate-permissions.ts
